+++ main/gameseg.c
@@ -111,11 +111,11 @@ int get_num_faces(side *sidep)
 }
 
 // Fill in array with four absolute point numbers for a given side
-void get_side_verts(short *vertlist,int segnum,int sidenum)
+void get_side_verts(int *vertlist,int segnum,int sidenum)
 {
 	int	i;
 	sbyte   *sv = Side_to_verts[sidenum];
-	short	*vp = Segments[segnum].verts;
+	int	*vp = Segments[segnum].verts;
 
 	for (i=4; i--;)
 		vertlist[i] = vp[sv[i]];
@@ -249,7 +249,7 @@ void create_all_vertnum_lists(int *num_faces, int *vertnums, int segnum, int sid
 // like create_all_vertex_lists(), but generate absolute point numbers
 void create_abs_vertex_lists(int *num_faces, int *vertices, int segnum, int sidenum, char *calling_file, int calling_linenum)
 {
-	short	*vp = Segments[segnum].verts;
+	int	*vp = Segments[segnum].verts;
 	side	*sidep = &Segments[segnum].sides[sidenum];
 	int  *sv = Side_to_verts_int[sidenum];
